for Nc::System: - Implementation of the cooperative visitor into the graph strucure - Integration of the cooperative visitor into the whole scene graph structure - Improvement of the class Utils::Clock
for Nc::Graphics:
- Creation of the class SpriteSheet to manage sprites with xml structure
- adding of a bunch of opengl enum

for GUI:
- Implementation of window style by using sprites sheet
- Adding a list of owned widget into a widget and implementation of a widget visitor
- Creation of Layout
- Creation of the basic widgets: ComboBox / LineEdit / ScrollBar / ScrollArea

Next step: Creation of the widget TextArea

Samples:
- Creation of a SimpleParticleEffect
- Creation of a SampleViewer (in development)

+++ Nc/Core/Math/Math.cpp
@@ -50,7 +50,7 @@ namespace Nc
             p[2] /= len;
         }
 
-        bool InRect(const Vector2f &RectPos, const Vector2f &RectSize, const Vector2f &pointTest)
+        bool InRect(const Vector2i &RectPos, const Vector2i &RectSize, const Vector2i &pointTest)
         {
             Vector2f posRectSize = RectPos + RectSize;
             if ((pointTest.Data[0] > RectPos.Data[0] && pointTest.Data[0] < posRectSize.Data[0] && posRectSize.Data[0] > RectPos.Data[0]) ||
